requirements.txt
slack-bolt 
apscheduler
numpy
torch
pandas
transformers
python-dotenv
pydrive
openai